+++ CHANGELOG.md
@@ -9,3 +9,5 @@ Major changes since the last busboy release (0.31):
 * Error on non-number limit rather than ignoring (#7)
 * Dicer is now part of the busboy itself and not an external dependency (#14)
 * Tests were converted to Mocha (#11, #12, #22, #23)
+* If preserveEmptyFilename is set to false, empty filenames will be changed to `undefined` (#52)
+* If preserveBlobFilename is set to false, `blob` as filename will be changed to `undefined` (#52)